backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-nvidia-ai-endpoints
python-dotenv
